Add confirmation pop up, add confirmation interface into Delete command

+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -10,6 +10,7 @@ import seedu.address.logic.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.person.Person;
+import seedu.address.ui.ConfirmationInterface;
 
 /**
  * Deletes a person identified using it's displayed index from the address book.
@@ -27,10 +28,14 @@ public class DeleteCommand extends Command {
 
     private final Index targetIndex;
 
-    public static final boolean confirmConfirmation = false;
+    public static final String MESSAGE_DELETE_CANCELLED =
+            "Delete cancelled.";
 
-    public DeleteCommand(Index targetIndex) {
+    private final ConfirmationInterface confirmationInterface;
+
+    public DeleteCommand(Index targetIndex, ConfirmationInterface confirmationInterface) {
         this.targetIndex = targetIndex;
+        this.confirmationInterface = confirmationInterface;
     }
 
     @Override
@@ -43,6 +48,14 @@ public class DeleteCommand extends Command {
         }
 
         Person personToDelete = lastShownList.get(targetIndex.getZeroBased());
+
+        // confirmation message
+        String displayMessage = Messages.format(personToDelete);
+        if (!confirmationInterface.confirm(
+                "Are you sure you want to delete " + displayMessage + "?")) {
+            return new CommandResult(MESSAGE_DELETE_CANCELLED);
+        }
+
         model.deletePerson(personToDelete);
         return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, Messages.format(personToDelete)));
     }
